Move transactionEstimatedSerializedSize to transactionvalidator
